migrate unit_modules_test_boto_dynamodb to pytest
